02_b_Mongo_DB/pyproject.toml
[tool.poetry]
name = "mongo-db"
version = "0.1.0"
description = ""
authors = ["Hassan Mehmood <ihassan463m@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.8"}
uvicorn = "^0.34.0"
pymongo = "^4.11"
passlib = "^1.7.4"
pyjwt = "^2.10.1"
python-jose = "^3.3.0"
bcrypt = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
